add support for the long-polling backend
this is the code we're actually running in production, though the
original version is made in coffee and being automatically compiled as
part of our asset compilation pipeline.

Because I didn't want to set up anything like this, the server included
here is the js-compiled version of the code.

The talk will be showing off the coffe version, but I'm sure you can
adapt accordingly.

This is also why I'm including underscore here and why I'm not using
jquery.Event - it allows me to reuse the existing code. For your
purpose, you might want to clean this up a bit